draft commit

+++ backend-app/etc/setup_django.sh
@@ -2,9 +2,14 @@
 sleep 10
 
 # prepare media files
+
 chmod 777 /media/
 cp -r /backend/media_files/* /media/
-rm -rf /backend/media_files
+
+if [[ $ENV = "production" ]]; then
+    rm -rf /backend/media_files
+fi
+
 
 # migrations
